v0.1/requirements.txt
Django==5.0.3
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-dotenv==1.0.1
django-allauth==0.61.1
django-admin-sortable2==2.1.10
stripe==7.12.0 